move fixture from test file to conftest.py test_auth_and_registration.py add start using redis
