+++ docker-compose.yml
@@ -29,5 +29,9 @@ services:
       - POSTGRES_DB=parser
     ports:
       - 5432:5432
+a  redis:
+    image: 'bitnami/redis:latest'
+    environment:
+      - ALLOW_EMPTY_PASSWORD=yes
 volumes:
   postgres_data:
