+++ requirements.txt
@@ -10,3 +10,5 @@ requests>=2.25.1, <3.0
 json-rpc>=1.13.0, <2.0
 pytest>=6.2.2, <7.0
 pytest-django>=4.1.0, <5.0
+django-redis==4.12.1
+redis == 3.5.3
