+++ tests/test_auth_and_registration.py
@@ -3,41 +3,6 @@ from rest_framework_simplejwt.backends import TokenBackend
 
 from api.models import CustomUser
 
-
-@pytest.fixture
-def add_test_user():
-    return CustomUser.objects.create_user(password='Password!!!',
-                                          first_name='TestUser',
-                                          last_name='TestUser',
-                                          email='email@email.email')
-
-
-@pytest.fixture
-def add_test_admin():
-    return CustomUser.objects.create_user(password='Password!!!',
-                                          first_name='TestUser',
-                                          last_name='TestUser',
-                                          email='admin@admin.admin',
-                                          is_superuser=True,
-                                          is_staff=True)
-
-
-@pytest.fixture
-def auth(client, add_test_user):
-    auth = client.post('/api/login',
-                       {'email': 'email@email.email',
-                        'password': 'Password!!!'})
-    return auth
-
-
-@pytest.fixture
-def auth_admin(client, add_test_user):
-    auth = client.post('/api/login',
-                       {'email': 'admin@admin.admin',
-                        'password': 'Password!!!'})
-    return auth
-
-
 @pytest.mark.django_db
 def test_register(client):
     response = client.post('/api/registration',
